Issue 32
Added a new id to the style.css file for the meet the team text style. Modified the footer to have the "meet the team" phrase link to the meet the team page. Modified app to support the meet the team page. Create the new file, MeettheTeam.jsx, to represent the five of us with our contact information and links to our biographies.